# packages/contracts-schema/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorBody": {
        "description": "统一错误体：``{\"error\": ...}`` 的内层字段。",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "additionalProperties": true,
            "title": "Details",
            "type": "object"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "request_id"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorResponse": {
        "description": "统一错误响应。",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SyncAckData": {
        "description": "同步确认结果：``already_acked`` 为 True 表示重复确认幂等成功。",
        "properties": {
          "already_acked": {
            "title": "Already Acked",
            "type": "boolean"
          },
          "envelope_id": {
            "title": "Envelope Id",
            "type": "string"
          }
        },
        "required": [
          "envelope_id",
          "already_acked"
        ],
        "title": "SyncAckData",
        "type": "object"
      },
      "SyncAckRequest": {
        "description": "同步信封确认请求（M3）：单个信封，应用成功后逐一确认。",
        "properties": {
          "envelope_id": {
            "title": "Envelope Id",
            "type": "string"
          }
        },
        "required": [
          "envelope_id"
        ],
        "title": "SyncAckRequest",
        "type": "object"
      },
      "SyncAckResponse": {
        "description": "``POST /sync/ack`` 响应。",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/SyncAckData"
          }
        },
        "required": [
          "data"
        ],
        "title": "SyncAckResponse",
        "type": "object"
      },
      "SyncEnvelopeData": {
        "description": "同步信封（密文中继）：云端只见密文，解密与校验在工作台完成。",
        "properties": {
          "ciphertext": {
            "title": "Ciphertext",
            "type": "string"
          },
          "created_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "envelope_id": {
            "title": "Envelope Id",
            "type": "string"
          },
          "event_id": {
            "title": "Event Id",
            "type": "string"
          },
          "expires_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "idempotency_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idempotency Key"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "target_device_id": {
            "title": "Target Device Id",
            "type": "string"
          }
        },
        "required": [
          "envelope_id",
          "event_id",
          "target_device_id",
          "ciphertext",
          "status"
        ],
        "title": "SyncEnvelopeData",
        "type": "object"
      },
...
      "SyncInjectRequest": {
        "description": "Mock 小程序事件注入请求（M3 dev 工具，生产环境禁用）。\n\n``payload`` 为模拟的小程序事件明文，服务端加密为密文信封后落库；\n``event_id`` 缺省由服务端生成，显式传入便于幂等重放联调。",
        "properties": {
          "event_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Event Id"
          },
          "idempotency_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idempotency Key"
          },
          "payload": {
            "additionalProperties": true,
            "title": "Payload",
            "type": "object"
          },
          "target_device_id": {
            "title": "Target Device Id",
            "type": "string"
          },
          "ttl_seconds": {
            "default": 604800,
            "maximum": 2592000.0,
            "minimum": 60.0,
            "title": "Ttl Seconds",
            "type": "integer"
          }
        },
        "required": [
          "target_device_id",
          "payload"
        ],
        "title": "SyncInjectRequest",
        "type": "object"
      },
      "SyncInjectResponse": {
        "description": "``POST /dev/sync/inject`` 响应（dev 工具）。",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/SyncEnvelopeData"
          }
        },
        "required": [
          "data"
        ],
        "title": "SyncInjectResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "云端控制平面 /api/v1：认证、账号、设备、状态流（M2）；心跳、配置、任务与小程序事件同步（M3）；技术日志与商户管理维持 stub（随开发者端页面交付）",
    "title": "warehouse-control-plane",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/dev/sync/inject": {
      "post": {
        "description": "dev 工具：Mock 小程序事件加密为密文信封写入 sync_envelope。\n\n仅供本地联调与端到端测试使用；生产环境一律 403 拒绝\n（不区分是否携带有效凭证，避免成为生产注入入口）。",
        "operationId": "inject_sync_event_api_v1_dev_sync_inject_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SyncInjectRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncInjectResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Mock 小程序事件注入（dev 工具）",
        "tags": [
          "sync"
        ]
      }
    },
...
    "/api/v1/sync/ack": {
      "post": {
        "description": "确认信封已应用：本地事务成功后逐个确认，重复 ACK 幂等成功。",
        "operationId": "ack_sync_api_v1_sync_ack_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SyncAckRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SyncAckResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "确认同步事件",
        "tags": [
          "sync"
        ]
      }
    },
...
  }
}